Create Excel to Premiere
In the example script provided earlier, the asset names and timecodes are expected to be formatted in a specific way in the Excel document. Specifically, the asset names are expected to be in the first column (column A), and the timecodes are expected to be in the second column (column B). The timecodes should be formatted as a string of timecode ranges, separated by commas. For example:

Asset Name	Timecodes
Footage 1	00:00:00:00-00:00:10:00,00:00:20:00-00:00:30:00
Footage 2	00:01:00:00-00:01:10:00,00:01:20:00-00:01:30:00